remove legacy paths
